Change local streaming to composition of camera thread inside camera streamer
